Complete Overhaul
I've completely rewritten the code and integrated a new locale. At present, I don’t see a need to expand its complexity or to separate it from the main file. Admittedly, I'm feeling a bit lazy about it. We can reassess and make further changes if needed in the future.