Locale update
Since my girlfriend isn't in the US, this had to be changed so it updates with it. Sorry to the countries that don't support it... But this is one way I could support a ton. You can request an issue, if enough are done I'll change it to be better :P